backend/requirements.txt
requests
mysql-connector-python
pandas
numpy
scikit-learn
pandas_ta
